feat: implement news list and update team flow
This commit introduces the `NewsList` widget to display learning content and implements the "update team" functionality.

Key changes:
- Added `NewsList` widget for displaying learning materials.
- Integrated `LearningListUseCase` to fetch learning content.
- Implemented `UpdateTeamUseCase` to allow users to modify their selected teams.
- Refactored `ProfileModel` to align with the latest API response.
- Updated `StockDataModel` and related components to use `TimeLeft` from `StatsModel`.
- Enhanced `StatsPage` with shimmer loading and improved data handling.
- Made UI improvements in `ContestDetailsPage` and `PersonalInfoPage`.
- Added internet permission to `AndroidManifest.xml`.
- General code refinements and bug fixes across various components related to contest management and data display.

+++ lib/utils/models/contest_model.dart
@@ -73,12 +73,12 @@ class Stock {
   bool isLiveData;
   int? livePoints;
   DateTime? lastUpdated;
-  int? selectionPercentage;
-  int? captainSelectionPercentage;
-  int? viceCaptainSelectionPercentage;
-  int? flexSelectionPercentage;
-  int? upPredictionPercentage;
-  int? downPredictionPercentage;
+  double? selectionPercentage;
+  double? captainSelectionPercentage;
+  double? viceCaptainSelectionPercentage;
+  double? flexSelectionPercentage;
+  double? upPredictionPercentage;
+  double? downPredictionPercentage;
   DateTime? createdAt;
   DateTime? updatedAt;
 
@@ -181,12 +181,12 @@ extension StockCopyWith on Stock {
     bool? isLiveData,
     int? livePoints,
     DateTime? lastUpdated,
-    int? selectionPercentage,
-    int? captainSelectionPercentage,
-    int? viceCaptainSelectionPercentage,
-    int? flexSelectionPercentage,
-    int? upPredictionPercentage,
-    int? downPredictionPercentage,
+    double? selectionPercentage,
+    double? captainSelectionPercentage,
+    double? viceCaptainSelectionPercentage,
+    double? flexSelectionPercentage,
+    double? upPredictionPercentage,
+    double? downPredictionPercentage,
   }) {
     return Stock(
       stockName: stockName ?? this.stockName,
